+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20180211182635) do
+ActiveRecord::Schema.define(version: 20180220211105) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -693,6 +693,16 @@ ActiveRecord::Schema.define(version: 20180211182635) do
   end
 
   add_index "moderators", ["user_id"], name: "index_moderators_on_user_id", using: :btree
+
+  create_table "newsletters", force: :cascade do |t|
+    t.string   "subject"
+    t.string   "segment_recipient", null: false
+    t.string   "from"
+    t.text     "body"
+    t.date     "sent_at"
+    t.datetime "created_at",        null: false
+    t.datetime "updated_at",        null: false
+  end
 
   create_table "notifications", force: :cascade do |t|
     t.integer  "user_id"